--- src/main/java/entity/WineEntity.java
@@ -1,35 +1,0 @@
-package entity;
-
-import org.jetbrains.annotations.NotNull;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class WineEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @NotNull
-    private String wineName;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getWineName() {
-        return wineName;
-    }
-
-    public void setWineName(String wineName) {
-        this.wineName = wineName;
-    }
-}
