+++ src/main/java/se/iths/corkdork/controller/CountryController.java
@@ -1,10 +1,10 @@
-package controller;
+package se.iths.corkdork.controller;
 
-import entity.CountryEntity;
+import se.iths.corkdork.entity.CountryEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import service.CountryService;
+import se.iths.corkdork.controller.service.CountryService;
 
 import java.util.Optional;
