+++ src/main/java/se/iths/corkdork/controller/UserController.java
@@ -1,10 +1,10 @@
-package controller;
+package se.iths.corkdork.controller;
 
-import entity.UserEntity;
+import se.iths.corkdork.entity.UserEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import service.UserService;
+import se.iths.corkdork.controller.service.UserService;
 
 import java.util.Optional;
