+++ src/main/java/se/iths/corkdork/entity/ColorEntity.java
@@ -1,4 +1,6 @@
-package entity;
+package se.iths.corkdork.entity;
+
+import org.jetbrains.annotations.NotNull;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,21 +8,21 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
-public class GrapeEntity {
+public class ColorEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private String name;
-    private String color;
 
+    @NotNull
+    private String color;
 
     public Long getId() {
         return id;
     }
 
-    public String getName() {
-        return name;
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public String getColor() {
